@@ -6,63 +6,65 @@
     "license": "MIT",
     "require": {
         "ext-json": "*",
-        "php": "^8.1",
-        "sylius/sylius": "1.12.*|1.13.*|1.14.*",
-        "3brs/sylius-shipment-export-plugin": "^v0.8.0"
+        "php": "^8.2",
+        "sylius/sylius": "^2.0",
+        "3brs/sylius-shipment-export-plugin": "SLS-5-sylius-2.0-dev"
     },
     "require-dev": {
-        "behat/behat": "^3.9.0",
-        "behat/mink-selenium2-driver": "^1.4",
-        "dmore/behat-chrome-extension": "^1.3",
-        "dmore/chrome-mink-driver": "^2.7",
-        "friends-of-behat/mink": "^1.10",
-        "friends-of-behat/mink-browserkit-driver": "^1.4",
-        "friends-of-behat/mink-debug-extension": "^2.1",
-        "friends-of-behat/mink-extension": "^2.4",
-        "friends-of-behat/page-object-extension": "^0.3",
-        "friends-of-behat/suite-settings-extension": "^1.0",
-        "friends-of-behat/symfony-extension": "^2.1",
-        "friends-of-behat/variadic-extension": "^1.3",
-        "php-http/message-factory": "^1.0.0",
-        "phpstan/phpstan": "^1.12",
-        "phpstan/phpstan-doctrine": "^1.5.7",
-        "phpstan/phpstan-strict-rules": "^1.3.0",
-        "phpstan/phpstan-symfony": "^1.3",
-        "phpstan/phpstan-webmozart-assert": "^1.2.0",
-        "phpunit/phpunit": "^9.5",
-        "polishsymfonycommunity/symfony-mocker-container": "^1.0",
-        "rector/rector": "^1.2",
-        "sylius-labs/coding-standard": "^4.2",
-        "symfony/browser-kit": "^6.4",
-        "symfony/debug-bundle": "^6.4",
-        "symfony/dotenv": "^6.4",
-        "symfony/flex": "^2.2.2",
-        "symfony/framework-bundle": "^6.4",
-        "symfony/http-foundation": "^6.4",
-        "symfony/http-kernel": "^6.4",
-        "symfony/intl": "^6.4",
-        "symfony/web-profiler-bundle": "^6.4",
-        "doctrine/annotations": "^2.0"
+        "behat/behat": "^v3.19.0",
+        "dmore/behat-chrome-extension": "^1.4.0",
+        "dmore/chrome-mink-driver": "^2.9.3",
+        "friends-of-behat/mink": "^v1.11.0",
+        "friends-of-behat/mink-browserkit-driver": "^v1.6.2",
+        "friends-of-behat/mink-debug-extension": "^v2.1.0",
+        "friends-of-behat/mink-extension": "^v2.7.5",
+        "friends-of-behat/page-object-extension": "^v0.3.2",
+        "friends-of-behat/suite-settings-extension": "^v1.1.0",
+        "friends-of-behat/symfony-extension": "v2.6.0",
+        "friends-of-behat/variadic-extension": "^v1.6.0",
+        "php-http/message-factory": "^1.1.0",
+        "phpstan/phpstan": "^2.1.10",
+        "phpstan/phpstan-doctrine": "^2.0.2",
+        "phpstan/phpstan-strict-rules": "^2.0.4",
+        "phpstan/phpstan-symfony": "^2.0.3",
+        "phpstan/phpstan-webmozart-assert": "^2.0.0",
+        "polishsymfonycommunity/symfony-mocker-container": "^v1.0.8",
+        "rector/rector": "^2.0.10",
+        "phpunit/phpunit": "^11.5",
+        "sylius-labs/coding-standard": "^v4.4.0",
+        "symfony/browser-kit": "^v6.4|^v7.1",
+        "symfony/debug-bundle": "^v6.4|^v7.1",
+        "symfony/dotenv": "^v6.4|^v7.1",
+        "symfony/flex": "^v2.5.0",
+        "symfony/framework-bundle": "^v6.4|^v7.1",
+        "symfony/http-foundation": "^v7.2.3",
+        "symfony/http-kernel": "^v6.4|^v7.1",
+        "symfony/intl": "^v6.4|^v7.1",
+        "symfony/web-profiler-bundle": "^v6.4|^v7.1",
+        "symplify/easy-coding-standard": "^12.0"
     },
     "conflict": {
+        "api-platform/symfony": "<v4.1.0",
         "api-platform/core": "<2.7.13",
         "babdev/pagerfanta-bundle": "<3.6.1",
         "doctrine/collections": "<1.7.0",
         "doctrine/data-fixtures": "<1.5.1",
         "doctrine/dbal": "<2.13.3",
-        "doctrine/doctrine-bundle": "<2.4.2",
+        "doctrine/doctrine-bundle": "<2.13.1",
         "friendsofsymfony/rest-bundle": "<3.1.0",
         "jms/serializer-bundle": "<4.2.0",
         "masterminds/html5": "<2.7.5",
         "knplabs/knp-menu": "<3.3.0",
         "lexik/jwt-authentication-bundle": "<2.12",
+        "payum/core": "<1.7.3",
         "polishsymfonycommunity/symfony-mocker-container": "<1.0.6",
         "sylius/resource-bundle": "<1.11.0",
+        "sylius/sylius": "<2.0.5",
         "symfony/css-selector": "<4.4.24",
         "symfony/framework-bundle": ">=5.4.0 <=5.4.20|>=6.0.0 <=6.0.16|>=6.1.0 <=6.1.8|>=6.2.0 <=6.2.2|6.2.8",
         "symfony/mime": "<5.4.0",
         "symfony/web-link": "<5.3.0",
-        "symplify/easy-coding-standard": "<10.2.0",
+        "symplify/easy-coding-standard": "<10.2.0|12.5.10",
         "twig/twig": "<2.14.7",
         "webmozart/assert": "<1.11.0",
         "willdurand/negotiation": "<3.0"
