@@ -4,44 +4,29 @@ declare(strict_types=1);
 
 namespace ThreeBRS\SyliusZasilkovnaPlugin\Entity;
 
+use Doctrine\DBAL\Types\Types;
 use Doctrine\ORM\Mapping as ORM;
 use Sylius\Component\Resource\Model\ResourceInterface;
 
-/**
- * @ORM\Table(name="threebrs_zasilkovna_config")
- *
- * @ORM\Entity
- */
 #[ORM\Table(name: 'threebrs_zasilkovna_config')]
 #[ORM\Entity]
 class ZasilkovnaConfig implements ResourceInterface, ZasilkovnaConfigInterface
 {
-    /**
-     * @ORM\Id
-     *
-     * @ORM\GeneratedValue(strategy="IDENTITY")
-     *
-     * @ORM\Column(type="integer", nullable=false)
-     */
     #[ORM\Id]
     #[ORM\GeneratedValue(strategy: 'IDENTITY')]
-    #[ORM\Column(type: 'integer', nullable: false)]
+    #[ORM\Column(type: Types::INTEGER, nullable: false)]
     protected ?int $id = null;
 
-    /** @ORM\Column(type="string", nullable=true) */
-    #[ORM\Column(type: 'string', nullable: true)]
+    #[ORM\Column(type: Types::STRING, nullable: true)]
     protected ?string $apiKey = null;
 
-    /** @ORM\Column(type="string", nullable=true) */
-    #[ORM\Column(type: 'string', nullable: true)]
+    #[ORM\Column(type: Types::STRING, nullable: true)]
     private ?string $senderLabel = null;
 
-    /** @ORM\Column(type="string", nullable=true) */
-    #[ORM\Column(type: 'string', nullable: true)]
+    #[ORM\Column(type: Types::STRING, nullable: true)]
     private ?string $carrierId = null;
 
-    /** @ORM\Column(type="string", nullable=true) */
-    #[ORM\Column(type: 'string', nullable: true)]
+    #[ORM\Column(type: Types::STRING, nullable: true)]
     private ?string $optionCountry = null;
 
     public function getId(): ?int
