@@ -5,16 +5,13 @@ declare(strict_types=1);
 namespace ThreeBRS\SyliusZasilkovnaPlugin\Model;
 
 use Doctrine\ORM\Mapping as ORM;
+use ThreeBRS\SyliusZasilkovnaPlugin\Entity\ZasilkovnaConfig;
 use ThreeBRS\SyliusZasilkovnaPlugin\Entity\ZasilkovnaConfigInterface;
 
 trait ZasilkovnaShippingMethodTrait
 {
-    /**
-     * @var ZasilkovnaConfigInterface|null
-     *
-     * @ORM\OneToOne(targetEntity="ThreeBRS\SyliusZasilkovnaPlugin\Entity\ZasilkovnaConfig", cascade={"persist"}, fetch="EAGER")
-     */
-    private $zasilkovnaConfig;
+    #[ORM\OneToOne(targetEntity: ZasilkovnaConfig::class, cascade: ['persist'], fetch: 'EAGER')]
+    private ?ZasilkovnaConfigInterface $zasilkovnaConfig = null;
 
     public function getZasilkovnaConfig(): ?ZasilkovnaConfigInterface
     {
