@@ -1,51 +1,48 @@
 const path = require('path');
 const Encore = require('@symfony/webpack-encore');
 
-const syliusBundles = path.resolve(__dirname, '../../vendor/sylius/sylius/src/Sylius/Bundle/');
-const uiBundleScripts = path.resolve(syliusBundles, 'UiBundle/Resources/private/js/');
-const uiBundleResources = path.resolve(syliusBundles, 'UiBundle/Resources/private/');
+const SyliusAdmin = require('@sylius-ui/admin');
+const SyliusShop = require('@sylius-ui/shop');
+
+// Admin config
+const adminConfig = SyliusAdmin.getWebpackConfig(path.resolve(__dirname));
 
 // Shop config
+const shopConfig = SyliusShop.getWebpackConfig(path.resolve(__dirname));
+
+// App shop config
 Encore
-  .setOutputPath('public/build/shop/')
-  .setPublicPath('/build/shop')
-  // .addEntry('shop-entry', './assets/shop/entry.js')
-  .addEntry('shop-entry', '../../vendor/sylius/sylius/src/Sylius/Bundle/ShopBundle/Resources/private/entry.js')
+  .setOutputPath('public/build/app/shop')
+  .setPublicPath('/build/app/shop')
+  .addEntry('app-shop-entry', './assets/shop/entrypoint.js')
   .disableSingleRuntimeChunk()
   .cleanupOutputBeforeBuild()
   .enableSourceMaps(!Encore.isProduction())
   .enableVersioning(Encore.isProduction())
-  .enableSassLoader();
+  .enableSassLoader()
+;
 
-const shopConfig = Encore.getWebpackConfig();
+const appShopConfig = Encore.getWebpackConfig();
 
-shopConfig.resolve.alias['sylius/ui'] = uiBundleScripts;
-shopConfig.resolve.alias['sylius/ui-resources'] = uiBundleResources;
-shopConfig.resolve.alias['sylius/bundle'] = syliusBundles;
-shopConfig.resolve.alias['chart.js/dist/Chart.min'] = path.resolve(__dirname, 'node_modules/chart.js/dist/chart.min.js');
-shopConfig.name = 'shop';
+appShopConfig.externals = Object.assign({}, appShopConfig.externals, { window: 'window', document: 'document' });
+appShopConfig.name = 'app.shop';
 
 Encore.reset();
 
-// Admin config
+// App admin config
 Encore
-  .setOutputPath('public/build/admin/')
-  .setPublicPath('/build/admin')
-  // .addEntry('admin-entry', './assets/admin/entry.js')
-  .addEntry('admin-entry', '../../vendor/sylius/sylius/src/Sylius/Bundle/AdminBundle/Resources/private/entry.js')
+  .setOutputPath('public/build/app/admin')
+  .setPublicPath('/build/app/admin')
+  .addEntry('app-admin-entry', './assets/admin/entrypoint.js')
   .disableSingleRuntimeChunk()
   .cleanupOutputBeforeBuild()
   .enableSourceMaps(!Encore.isProduction())
   .enableVersioning(Encore.isProduction())
   .enableSassLoader();
 
-const adminConfig = Encore.getWebpackConfig();
+const appAdminConfig = Encore.getWebpackConfig();
 
-adminConfig.resolve.alias['sylius/ui'] = uiBundleScripts;
-adminConfig.resolve.alias['sylius/ui-resources'] = uiBundleResources;
-adminConfig.resolve.alias['sylius/bundle'] = syliusBundles;
-adminConfig.resolve.alias['chart.js/dist/Chart.min'] = path.resolve(__dirname, 'node_modules/chart.js/dist/chart.min.js');
-adminConfig.externals = Object.assign({}, adminConfig.externals, { window: 'window', document: 'document' });
-adminConfig.name = 'admin';
+appAdminConfig.externals = Object.assign({}, appAdminConfig.externals, { window: 'window', document: 'document' });
+appAdminConfig.name = 'app.admin';
 
-module.exports = [shopConfig, adminConfig];
+module.exports = [shopConfig, adminConfig, appShopConfig, appAdminConfig];
